Initial router vault swap test + router openPosition tests to WasabiRouter.test.ts
